+++ src/transpiler/tests/csout/test45a.cs
@@ -29,7 +29,7 @@ public static partial class Program
 {
     public static void LoadFlags(string cFile, ref dynamic[] aFlag, decimal nCount = default)
     {
-        int i = default;
+        long i = default;
 
         HbRuntime.ASize(ref aFlag, nCount);
         for (i = 1; i <= nCount; i++)
